Rename media queries file
